Version: 2011-01-10 ksh93u
Sourced from ast-open-history version 2011-01-10

+++ src/lib/libast/vmalloc/vmmopen.c
@@ -1,10 +1,10 @@
 /***********************************************************************
 *                                                                      *
 *               This software is part of the ast package               *
-*                  Copyright (c) 1985-2005 AT&T Corp.                  *
+*          Copyright (c) 1985-2011 AT&T Intellectual Property          *
 *                      and is licensed under the                       *
 *                  Common Public License, Version 1.0                  *
-*                            by AT&T Corp.                             *
+*                    by AT&T Intellectual Property                     *
 *                                                                      *
 *                A copy of the License is available at                 *
 *            http://www.opensource.org/licenses/cpl1.0.txt             *
@@ -121,7 +121,7 @@ Mmvm_t*	mm;	/* to return some mapped info	*/
 			goto done;
 
 	new_f:	/* create an initial set of data */
-		size = round;
+		size = (off_t)round;
 		if(lseek(fd, size-1, 0) != (size-1) || write(fd, "", 1) != 1 )
 			goto done;
