+++ src/lib/libast/vmalloc/vmpool.c
@@ -1,7 +1,7 @@
 /***********************************************************************
 *                                                                      *
 *               This software is part of the ast package               *
-*          Copyright (c) 1985-2010 AT&T Intellectual Property          *
+*          Copyright (c) 1985-2011 AT&T Intellectual Property          *
 *                      and is licensed under the                       *
 *                  Common Public License, Version 1.0                  *
 *                    by AT&T Intellectual Property                     *
@@ -165,7 +165,7 @@ reg Void_t*	addr;
 				if(bp == tp)
 					goto done;
 
-		offset = (Vmuchar_t*)addr - (Vmuchar_t*)tp;
+		offset = (long)((Vmuchar_t*)addr - (Vmuchar_t*)tp);
 		goto done;
 	}
